pyproject.toml
[tool.poetry]
name = "djankiserv"
version = "0.0.2"
description = "Django-based synchronisation and API server for Anki"
authors = ["Anton Melser <anton@melser.org>"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "~3.8"
django = "^3.1"
djangorestframework = "^3.11.1"
gunicorn = "^20.0.4"
django-k8s = "^0.2.9"
djangorestframework-simplejwt = "^4.4.0"
psycopg2-binary = {version = "^2.8.6", optional = true}
mysqlclient = {version = "^2.0.1", optional = true}

[tool.poetry.dev-dependencies]
requests-mock = "^1.8.0"
coverage = {extras = ["toml"], version = "^5.2.1"}
python-coveralls = "^2.9.3"
coveralls = "^2.1.2"
colorama = "^0.4.3"
pylint = "^2.6.0"
flake9 = "^3.8.3"
pre-commit = "^2.7.1"
mysqlclient = "^2.0.1"
psycopg2-binary = "^2.8.6"

[tool.coverage.run]
omit = [
    "*manage.py",
]

[tool.pylint.messages_control]
# I0011 Warning locally suppressed using disable-msg
disable = """
    invalid-name,
    logging-fstring-interpolation,
    fixme,
    I0011,
    missing-docstring,
    C0330,
    too-few-public-methods,
    too-many-arguments,
    C0326,
    pointless-string-statement,
    no-member,
    R0901,
    R0801,
"""

ignore = """,
    migrations,
"""

[tool.pylint.basic]
# Good variable names which should always be accepted, separated by a comma.
good-names = "i,j,k,ex,_,f,e,s,w,li,t"

[tool.pylint.format]
max-line-length = 120

[tool.isort]
multi_line_output = 3
line_length = 120
known_third_party = "rest_framework"
known_first_party = "djankiserv"
include_trailing_comma = "True"
force_grid_wrap = 0
use_parentheses = "True"

[tool.flake8]
ignore = "E226,E266"
exclude = """,
    tmp,
    libs,
    .git,
    __pycache__,
    build,
    dist
"""
max-line-length = 120
max-complexity = 10
extend-ignore = "E203,W503"

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    # Add additional patterns here:
    | .ropeproject
    | tmp
    | libs
    | static
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
